+++ es-app/src/scrapers/Scraper.cpp
@@ -29,7 +29,8 @@ std::unique_ptr<ScraperSearchHandle> startScraperSearch(const ScraperSearchParam
 
     // Check if the scraper in the settings still exists as a registered scraping source.
     if (scraper_request_funcs.find(name) == scraper_request_funcs.end())
-        LOG(LogWarning) << "Configured scraper (" << name << ") unavailable, scraping aborted.";
+        LOG(LogWarning) << "Warning - Configured scraper (" << name <<
+                ") unavailable, scraping aborted.";
     else
         scraper_request_funcs.at(name)(params, handle->mRequestQueue, handle->mResults);
 
@@ -44,7 +45,8 @@ std::unique_ptr<ScraperSearchHandle> startMediaURLsFetch(const std::string& game
     ScraperSearchParams params;
     // Check if the scraper in the settings still exists as a registered scraping source.
     if (scraper_request_funcs.find(name) == scraper_request_funcs.end())
-        LOG(LogWarning) << "Configured scraper (" << name << ") unavailable, scraping aborted.";
+        LOG(LogWarning) << "Warning - Configured scraper (" << name <<
+                ") unavailable, scraping aborted.";
     else
         // Specifically use the TheGamesDB function as this type of request
         // will never occur for ScreenScraper.
@@ -140,7 +142,7 @@ void ScraperHttpRequest::update()
         return;
 
     // Everything else is some sort of error.
-    LOG(LogError) << "ScraperHttpRequest network error (status: " << status<< ") - "
+    LOG(LogError) << "Error - ScraperHttpRequest network error (status: " << status<< ") - "
             << mReq->getErrorMsg();
     setError(mReq->getErrorMsg());
 }
@@ -367,7 +369,7 @@ bool resizeImage(const std::string& path, int maxWidth, int maxHeight)
     if (format == FIF_UNKNOWN)
         format = FreeImage_GetFIFFromFilename(path.c_str());
     if (format == FIF_UNKNOWN) {
-        LOG(LogError) << "Error - could not detect filetype for image \"" << path << "\"!";
+        LOG(LogError) << "Error - Could not detect filetype for image \"" << path << "\"!";
         return false;
     }
 
@@ -376,7 +378,7 @@ bool resizeImage(const std::string& path, int maxWidth, int maxHeight)
         image = FreeImage_Load(format, path.c_str());
     }
     else {
-        LOG(LogError) << "Error - file format reading not supported for image \"" << path << "\"!";
+        LOG(LogError) << "Error - File format not supported for image \"" << path << "\"!";
         return false;
     }
 
@@ -397,7 +399,7 @@ bool resizeImage(const std::string& path, int maxWidth, int maxHeight)
     FreeImage_Unload(image);
 
     if (imageRescaled == nullptr) {
-        LOG(LogError) << "Could not resize image! (not enough memory? invalid bitdepth?)";
+        LOG(LogError) << "Error - Could not resize image! (not enough memory? invalid bitdepth?)";
         return false;
     }
 
@@ -405,7 +407,7 @@ bool resizeImage(const std::string& path, int maxWidth, int maxHeight)
     FreeImage_Unload(imageRescaled);
 
     if (!saved) {
-        LOG(LogError) << "Failed to save resized image!";
+        LOG(LogError) << "Error - Failed to save resized image!";
     }
 
     return saved;