+++ es-core/src/resources/ResourceManager.cpp
@@ -8,6 +8,10 @@
 #include "ResourceManager.h"
 
 #include "utils/FileSystemUtil.h"
+#include "Log.h"
+#include "Platform.h"
+#include "Scripting.h"
+
 #include <fstream>
 
 auto array_deleter = [](unsigned char* p) { delete[] p; };
@@ -31,23 +35,36 @@ std::string ResourceManager::getResourcePath(const std::string& path) const
 {
     // Check if this is a resource file.
     if ((path[0] == ':') && (path[1] == '/')) {
-        std::string test;
+        std::string testHome;
+        std::string testDataPath;
 
         // Check under the home directory.
-        test = Utils::FileSystem::getHomePath() + "/.emulationstation/resources/" + &path[2];
-        if (Utils::FileSystem::exists(test))
-            return test;
+        testHome = Utils::FileSystem::getHomePath() + "/.emulationstation/resources/" + &path[2];
+        if (Utils::FileSystem::exists(testHome))
+            return testHome;
 
         // Check for the resource under the data installation directory for Unix or under
         // the executable directory for Windows.
         #ifdef _WIN64
-        test = Utils::FileSystem::getExePath() + "/resources/" + &path[2];
+        testDataPath = Utils::FileSystem::getExePath() + "/resources/" + &path[2];
         #else
-        test = Utils::FileSystem::getProgramDataPath() + "/resources/" + &path[2];
+        testDataPath = Utils::FileSystem::getProgramDataPath() + "/resources/" + &path[2];
         #endif
 
-        if (Utils::FileSystem::exists(test))
-        return test;
+        if (Utils::FileSystem::exists(testDataPath)) {
+            return testDataPath;
+        }
+        // For missing resources, log an error and terminate the application. This should
+        // indicate that we have a broken EmulationStation installation.
+        else {
+            LOG(LogError) << "Error - Program resource missing: " << path;
+            LOG(LogError) << "Tried to find the resource in the following locations:";
+            LOG(LogError) << testHome;
+            LOG(LogError) << testDataPath;
+            LOG(LogError) << "Has EmulationStation been properly installed?";
+            Scripting::fireEvent("quit");
+            emergencyShutdown();
+        }
     }
 
     // Not a resource, return unmodified path.